Added project
The specifications for this project where: Build a checkout form (HTML & CSS only) in Bootstrap. Research the classic fields. Use fields with at least these types:
* Text,
* Select,
* password,
* e-mail,
* checkbox,
* radio button,
* Textarea.
Note the labels for the fields and frames for the individual sections in the form.
Use at least the following fields:
* Salutation,
* First name and surname,
* age/date of birth,
* Delivery address with the various fields,
* Payment method,
* Consent to a fictitious privacy policy.

Additionally:
Insert the bootstrap navigation.
As destinations in the navigation, there should be a link to two other HTML pages (e.g. the CV or the burger example from the previous semester).